+++ src/.idea/.idea.OpenAC.Net.Devices/.idea/.name
@@ -1,0 +1,1 @@
+OpenAC.Net.Devices
